requirements.txt
Pillow
matplotlib
networkx
streamlit
pyvis